+++ sandboxes/common/test/lib/input-json-for-run-compiled.bash
@@ -2,21 +2,19 @@
 #inline var-is-set.bash
 #inline var-not-empty.bash
 
-function compiled_program {
+function compiled_program_json {
   set -e
   fun_is_set sandbox
   var_not_empty COMPILE_IMAGE
   var_not_empty CONTAINER_NAME
   var_is_set source_program
-  echo '{}' | \
-      jshon -s "$source_program" -i sourceProgram | \
-      IMAGE=$COMPILE_IMAGE CONTAINER_NAME=compile-$CONTAINER_NAME sandbox | \
-      jshon -e compiledProgram -u
+  jq -n --arg sp "$source_program" '{sourceProgram: $sp}' | \
+    IMAGE=$COMPILE_IMAGE CONTAINER_NAME=compile-$CONTAINER_NAME sandbox | \
+    jq '{"compiledProgram"}'
 }
 
 function input_json {
   set -e
   : ${stdin:=''}
-  echo "{ \"compiledProgram\": \"$(compiled_program)\" }" | \
-      jshon -s "$stdin" -i stdin
+  compiled_program_json | jq --arg stdin "$stdin" '. "stdin" = $stdin'
 }