+++ sandboxes/common/test/lib/input-json-for-run-interpreted.bash
@@ -4,7 +4,8 @@ function input_json {
   set -e
   var_is_set interpreted_program
   : ${stdin:=''}
-  echo '{}' | \
-    jshon -s "$stdin" -i stdin | \
-    jshon -s "$interpreted_program" -i interpretedProgram
+  jq -n \
+    --arg stdin "$stdin" \
+    --arg ip "$interpreted_program" \
+    '{stdin: $stdin, interpretedProgram: $ip}'
 }